refactor: cleaner project structure
